feat: web service for semantics description
